Back-end/requirements.txt
flask
flask-socketio
flask-jwt-extended
pymongo
python-dotenv
bcrypt
flask_cors